refactor(settings): extract theme page from the settings layer
Move the theme catalog, search, and editor into a dedicated module so the
settings layer is no longer responsible for that page's construction.
