+++ src/ui/settings.rs
@@ -20,7 +20,9 @@ use crate::{
 };
 
 mod bindings;
+mod theme;
 use bindings::{bind_choice, bind_switch};
+use theme::theme_page;
 
 #[cfg(test)]
 mod tests;
@@ -29,9 +31,9 @@ use super::{
     blur::BlurBin,
     browser::{dismiss_modal_layer, modal_layer},
     browser_modes::{BrowserMode, ClickActivation, ClickCount},
-    controls::{form_entry, menu_option, modal_layout, segmented_control},
+    controls::{menu_option, modal_layout, segmented_control},
     terminal,
-    theme::{TextSize, Theme, ThemeManager, ThemeTokens},
+    theme::ThemeManager,
 };
 
 pub(super) type UpdateNoticeHandler = Rc<dyn Fn(Option<(ReleaseMetadata, String, UpdateMethod)>)>;
@@ -2435,542 +2437,6 @@ fn append_keybinding(content: &gtk::Box, label: &str, keys: &str) {
     content.append(&row);
 }
 
-fn theme_page(manager: Rc<ThemeManager>) -> (gtk::Widget, Vec<(gtk::FlowBox, u32)>) {
-    let content = page_content();
-    content.add_css_class("theme-page");
-
-    let follow = gtk::Switch::builder()
-        .active(manager.follows_omarchy())
-        .valign(gtk::Align::Center)
-        .build();
-    if manager.is_omarchy_available() {
-        append_heading(&content, "SYSTEM");
-        let system = gtk::Box::new(gtk::Orientation::Horizontal, 14);
-        system.add_css_class("settings-option");
-        let icon = crate::assets::primary_icon(icons::MONITOR, 22);
-        icon.add_css_class("system-theme-icon");
-        let copy = gtk::Box::new(gtk::Orientation::Vertical, 2);
-        copy.set_hexpand(true);
-        copy.set_valign(gtk::Align::Center);
-        let system_title = gtk::Label::new(Some("Follow Omarchy"));
-        system_title.set_xalign(0.0);
-        system_title.add_css_class("settings-option-title");
-        let system_description = gtk::Label::new(Some(
-            "Use the active Omarchy Quattro theme and follow system theme changes.",
-        ));
-        system_description.set_xalign(0.0);
-        system_description.set_wrap(true);
-        system_description.add_css_class("settings-option-description");
-        copy.append(&system_title);
-        copy.append(&system_description);
-        system.append(&icon);
-        system.append(&copy);
-        system.append(&follow);
-        content.append(&system);
-    }
-
-    append_heading(&content, "TYPOGRAPHY");
-    let text_sizes = [TextSize::Small, TextSize::Medium, TextSize::Large];
-    let active_text_size = text_sizes
-        .iter()
-        .position(|&size| size == manager.text_size())
-        .unwrap_or(1);
-    let (text_size_control, text_size_buttons) =
-        segmented_control(&["Small", "Medium", "Large"], active_text_size);
-    let text_size_row = gtk::Box::new(gtk::Orientation::Vertical, 8);
-    text_size_row.add_css_class("settings-option");
-    let text_size_copy = gtk::Box::new(gtk::Orientation::Vertical, 2);
-    text_size_copy.set_hexpand(true);
-    let text_size_title = gtk::Label::new(Some("Text size"));
-    text_size_title.set_xalign(0.0);
-    text_size_title.add_css_class("settings-option-title");
-    let text_size_description = gtk::Label::new(Some(
-        "Scale interface text across menus, labels, and lists.",
-    ));
-    text_size_description.set_xalign(0.0);
-    text_size_description.set_wrap(true);
-    text_size_description.add_css_class("settings-option-description");
-    text_size_copy.append(&text_size_title);
-    text_size_copy.append(&text_size_description);
-    text_size_row.append(&text_size_copy);
-    text_size_row.append(&text_size_control);
-    content.append(&text_size_row);
-    for (button, size) in text_size_buttons.into_iter().zip(text_sizes) {
-        bind_choice(
-            &manager,
-            &button,
-            size,
-            ThemeManager::text_size,
-            ThemeManager::set_text_size,
-        );
-    }
-
-    append_heading(&content, "THEMES");
-    let packaged = gtk::FlowBox::builder()
-        .column_spacing(12)
-        .row_spacing(12)
-        .max_children_per_line(3)
-        .min_children_per_line(1)
-        .selection_mode(gtk::SelectionMode::None)
-        .homogeneous(true)
-        .build();
-    packaged.add_css_class("theme-grid");
-    let theme_search = gtk::Entry::new();
-    theme_search.add_css_class("form-control");
-    theme_search.add_css_class("theme-search");
-    theme_search.set_placeholder_text(Some("Search themes"));
-    let search_keys = gtk::EventControllerKey::new();
-    search_keys.set_propagation_phase(gtk::PropagationPhase::Capture);
-    let selected_search = theme_search.downgrade();
-    search_keys.connect_key_pressed(move |_, key, _, modifiers| {
-        if modifiers.contains(gdk::ModifierType::CONTROL_MASK)
-            && matches!(key, gdk::Key::a | gdk::Key::A)
-            && let Some(search) = selected_search.upgrade()
-        {
-            search.select_region(0, -1);
-            return glib::Propagation::Stop;
-        }
-        glib::Propagation::Proceed
-    });
-    theme_search.add_controller(search_keys);
-    let clear_search = gtk::Button::builder()
-        .child(&crate::assets::primary_icon(icons::X, 15))
-        .tooltip_text("Clear theme search")
-        .halign(gtk::Align::End)
-        .valign(gtk::Align::Center)
-        .margin_end(6)
-        .visible(false)
-        .build();
-    clear_search.add_css_class("theme-search-clear");
-    clear_search.set_has_frame(false);
-    let search_overlay = gtk::Overlay::new();
-    search_overlay.set_child(Some(&theme_search));
-    search_overlay.add_overlay(&clear_search);
-    content.append(&search_overlay);
-    let cleared_search = theme_search.clone();
-    clear_search.connect_clicked(move |_| {
-        cleared_search.set_text("");
-        cleared_search.grab_focus();
-    });
-    let (appearance_filter, appearance_buttons) = segmented_control(&["All", "Light", "Dark"], 0);
-    appearance_filter.add_css_class("theme-appearance-filter");
-    content.append(&appearance_filter);
-    let catalog_scroll = gtk::ScrolledWindow::builder()
-        .child(&packaged)
-        .hscrollbar_policy(gtk::PolicyType::Never)
-        .vscrollbar_policy(gtk::PolicyType::Automatic)
-        .min_content_height(240)
-        .max_content_height(330)
-        .propagate_natural_height(true)
-        .build();
-    catalog_scroll.add_css_class("theme-catalog-scroll");
-    let catalog_container = gtk::Box::new(gtk::Orientation::Vertical, 0);
-    catalog_container.add_css_class("theme-catalog-container");
-    catalog_container.append(&catalog_scroll);
-    content.append(&catalog_container);
-
-    append_heading(&content, "YOUR THEMES");
-    let custom = gtk::FlowBox::builder()
-        .column_spacing(12)
-        .row_spacing(12)
-        .max_children_per_line(3)
-        .min_children_per_line(1)
-        .selection_mode(gtk::SelectionMode::None)
-        .homogeneous(true)
-        .build();
-    custom.add_css_class("theme-grid");
-    content.append(&custom);
-
-    let mut catalog_cards = Vec::new();
-    for theme in manager.themes() {
-        let custom_theme = theme.custom;
-        let name = theme.tokens.name.clone();
-        let light = theme_is_light(&theme.tokens);
-        let flow = if custom_theme { &custom } else { &packaged };
-        let child = append_theme_card(flow, theme, &manager);
-        if !custom_theme {
-            catalog_cards.push((child, name, light));
-        }
-    }
-    let catalog_cards = Rc::new(catalog_cards);
-    let appearance = Rc::new(Cell::new(ThemeAppearance::All));
-    let filtered_cards = catalog_cards.clone();
-    let filtered_appearance = appearance.clone();
-    let filter_search = theme_search.clone();
-    let apply_catalog_filter: Rc<dyn Fn()> = Rc::new(move || {
-        let query = filter_search.text();
-        let appearance = filtered_appearance.get();
-        for (child, name, light) in filtered_cards.iter() {
-            let appearance_matches = match appearance {
-                ThemeAppearance::All => true,
-                ThemeAppearance::Dark => !light,
-                ThemeAppearance::Light => *light,
-            };
-            child.set_visible(appearance_matches && theme_name_matches(name, &query));
-        }
-    });
-    let search_filter = apply_catalog_filter.clone();
-    theme_search.connect_changed(move |search| {
-        clear_search.set_visible(!search.text().is_empty());
-        search_filter();
-    });
-    for (button, value) in appearance_buttons.into_iter().zip([
-        ThemeAppearance::All,
-        ThemeAppearance::Light,
-        ThemeAppearance::Dark,
-    ]) {
-        let appearance = appearance.clone();
-        let apply_filter = apply_catalog_filter.clone();
-        button.connect_toggled(move |button| {
-            if button.is_active() {
-                appearance.set(value);
-                apply_filter();
-            }
-        });
-    }
-
-    let add = gtk::Button::new();
-    add.add_css_class("add-theme-card");
-    add.set_has_frame(false);
-    let add_content = gtk::Box::new(gtk::Orientation::Vertical, 7);
-    add_content.set_halign(gtk::Align::Center);
-    add_content.set_valign(gtk::Align::Center);
-    let plus = crate::assets::primary_icon(icons::PLUS, 22);
-    let add_label = gtk::Label::new(Some("Add a theme"));
-    add_content.append(&plus);
-    add_content.append(&add_label);
-    add.set_child(Some(&add_content));
-    custom.insert(&add, -1);
-
-    let known_custom = RefCell::new(
-        manager
-            .themes()
-            .into_iter()
-            .filter(|theme| theme.custom)
-            .map(|theme| theme.id)
-            .collect::<std::collections::HashSet<_>>(),
-    );
-    let weak_manager = Rc::downgrade(&manager);
-    manager.bind_preference(
-        &custom,
-        |manager| {
-            manager
-                .themes()
-                .into_iter()
-                .filter(|theme| theme.custom)
-                .map(|theme| theme.id)
-                .collect::<Vec<_>>()
-        },
-        move |widget, _| {
-            let Some(flow) = widget.downcast_ref::<gtk::FlowBox>() else {
-                return;
-            };
-            if let Some(manager) = weak_manager.upgrade() {
-                for theme in manager.themes().into_iter().filter(|theme| theme.custom) {
-                    if known_custom.borrow_mut().insert(theme.id.clone()) {
-                        append_theme_card(flow, theme, &manager);
-                    }
-                }
-            }
-        },
-    );
-    let (editor, editor_fields) = theme_editor(manager.clone());
-    editor.set_reveal_child(false);
-    content.append(&editor);
-    let shown_editor = editor.clone();
-    add.connect_clicked(move |_| shown_editor.set_reveal_child(true));
-
-    let scroller = scrollable_page(&content, None);
-    bind_switch(
-        &manager,
-        &follow,
-        ThemeManager::follows_omarchy,
-        ThemeManager::set_follow_omarchy,
-    );
-    (
-        scroller,
-        vec![(packaged, 3), (custom, 3), (editor_fields, 4)],
-    )
-}
-
-#[derive(Clone, Copy)]
-enum ThemeAppearance {
-    All,
-    Dark,
-    Light,
-}
-
-fn theme_name_matches(name: &str, query: &str) -> bool {
-    let query = query.trim().to_lowercase();
-    query.is_empty() || name.to_lowercase().contains(&query)
-}
-
-fn theme_is_light(tokens: &ThemeTokens) -> bool {
-    theme_background_is_light(&tokens.background)
-}
-
-fn theme_background_is_light(background: &str) -> bool {
-    let value = background.strip_prefix('#').unwrap_or_default();
-    let Ok(color) = u32::from_str_radix(value, 16) else {
-        return false;
-    };
-    let channel = |shift| {
-        let value = f64::from((color >> shift) & 0xff_u32) / 255.0;
-        if value <= 0.04045 {
-            value / 12.92
-        } else {
-            ((value + 0.055) / 1.055).powf(2.4)
-        }
-    };
-    let luminance = 0.2126 * channel(16) + 0.7152 * channel(8) + 0.0722 * channel(0);
-    luminance > 0.4
-}
-
-fn append_theme_card(
-    flow: &gtk::FlowBox,
-    theme: Theme,
-    manager: &Rc<ThemeManager>,
-) -> gtk::FlowBoxChild {
-    let card = gtk::Button::new();
-    card.add_css_class("theme-card");
-    card.set_has_frame(false);
-    card.set_overflow(gtk::Overflow::Visible);
-    let content = gtk::Box::new(gtk::Orientation::Vertical, 6);
-    let preview = gtk::Overlay::new();
-    preview.set_child(Some(&theme_preview(&theme.tokens)));
-    let check = gtk::Image::from_icon_name(icons::CHECK_ON_PRIMARY);
-    check.add_css_class("theme-card-check");
-    check.set_halign(gtk::Align::End);
-    check.set_valign(gtk::Align::Start);
-    check.set_margin_top(8);
-    check.set_margin_end(8);
-    check.set_pixel_size(10);
-    preview.add_overlay(&check);
-    content.append(&preview);
-    let label_row = gtk::Box::new(gtk::Orientation::Horizontal, 7);
-    let selected = !manager.follows_omarchy() && manager.selected_id() == theme.id;
-    check.set_visible(selected);
-    if selected {
-        card.add_css_class("selected");
-    }
-    let label = gtk::Label::new(Some(&theme.tokens.name));
-    label.set_xalign(0.0);
-    label.set_ellipsize(gtk::pango::EllipsizeMode::End);
-    label_row.append(&label);
-    content.append(&label_row);
-    card.set_child(Some(&content));
-    let theme_id = theme.id;
-    let selected_theme = theme_id.clone();
-    let check = check.downgrade();
-    manager.bind_preference(
-        &card,
-        move |manager| !manager.follows_omarchy() && manager.selected_id() == selected_theme,
-        move |card, selected| {
-            if selected {
-                card.add_css_class("selected");
-            } else {
-                card.remove_css_class("selected");
-            }
-            if let Some(check) = check.upgrade() {
-                check.set_visible(selected);
-            }
-        },
-    );
-    let manager = manager.clone();
-    card.connect_clicked(move |_| {
-        manager.select_theme(&theme_id);
-    });
-    flow.insert(&card, -1);
-    card.parent()
-        .and_downcast::<gtk::FlowBoxChild>()
-        .expect("FlowBox must wrap inserted theme cards")
-}
-
-fn theme_preview(tokens: &ThemeTokens) -> gtk::DrawingArea {
-    let area = gtk::DrawingArea::new();
-    area.add_css_class("theme-preview");
-    area.set_content_width(190);
-    area.set_content_height(72);
-    let tokens = tokens.clone();
-    area.set_draw_func(move |_, context, width, height| {
-        let color = |value: &str| gdk::RGBA::parse(value).unwrap_or(gdk::RGBA::BLACK);
-        let paint = |context: &gtk::cairo::Context, value: &str| {
-            let value = color(value);
-            context.set_source_rgba(
-                f64::from(value.red()),
-                f64::from(value.green()),
-                f64::from(value.blue()),
-                1.0,
-            );
-        };
-        context.rounded_rectangle(0.0, 0.0, f64::from(width), f64::from(height), 6.0);
-        context.clip();
-        paint(context, &tokens.background);
-        context.rectangle(0.0, 0.0, f64::from(width), f64::from(height));
-        let _ = context.fill();
-        paint(context, &tokens.surface);
-        context.rectangle(0.0, 0.0, f64::from(width) * 0.40, f64::from(height));
-        let _ = context.fill();
-        for (x, y, w, value) in [
-            (10.0, 23.0, 45.0, &tokens.dim_text),
-            (10.0, 36.0, 59.0, &tokens.accent),
-            (10.0, 51.0, 39.0, &tokens.dim_text),
-            (f64::from(width) * 0.45, 23.0, 47.0, &tokens.accent),
-            (f64::from(width) * 0.45, 37.0, 83.0, &tokens.dim_text),
-            (f64::from(width) * 0.45, 51.0, 66.0, &tokens.dim_text),
-        ] {
-            paint(context, value);
-            context.rounded_rectangle(x, y, w, 5.0, 2.5);
-            let _ = context.fill();
-        }
-    });
-    area
-}
-
-fn theme_editor(manager: Rc<ThemeManager>) -> (gtk::Revealer, gtk::FlowBox) {
-    let panel = gtk::Box::new(gtk::Orientation::Vertical, 12);
-    panel.add_css_class("theme-editor");
-    let header = gtk::Box::new(gtk::Orientation::Horizontal, 8);
-    let title = gtk::Label::new(Some("Add a theme"));
-    title.add_css_class("settings-option-title");
-    title.set_xalign(0.0);
-    title.set_hexpand(true);
-    header.append(&title);
-    panel.append(&header);
-    let name = form_entry();
-    name.set_placeholder_text(Some("Theme name"));
-    panel.append(&name);
-
-    let values = Rc::new(RefCell::new(manager.starter_tokens()));
-    let fields = gtk::FlowBox::builder()
-        .column_spacing(18)
-        .row_spacing(10)
-        .max_children_per_line(4)
-        .min_children_per_line(1)
-        .selection_mode(gtk::SelectionMode::None)
-        .homogeneous(true)
-        .build();
-    fields.add_css_class("theme-color-fields");
-    for (label_text, field) in [
-        ("Background", ColorField::Background),
-        ("Surface", ColorField::Surface),
-        ("Text", ColorField::Text),
-        ("Accent", ColorField::Accent),
-        ("Danger", ColorField::Danger),
-        ("Muted", ColorField::Muted),
-        ("Highlight", ColorField::Highlight),
-        ("Border", ColorField::Border),
-        ("Dim text", ColorField::DimText),
-    ] {
-        let field_row = gtk::Box::new(gtk::Orientation::Horizontal, 7);
-        let label = gtk::Label::new(Some(label_text));
-        label.set_xalign(0.0);
-        let dialog = gtk::ColorDialog::builder()
-            .title(format!("Choose {label_text}"))
-            .with_alpha(false)
-            .build();
-        let picker = gtk::ColorDialogButton::new(Some(dialog));
-        picker.add_css_class("theme-color-picker");
-        if let Ok(color) = gdk::RGBA::parse(field.get(&values.borrow())) {
-            picker.set_rgba(&color);
-        }
-        let values_for_color = values.clone();
-        let manager_for_color = manager.clone();
-        picker.connect_rgba_notify(move |picker| {
-            field.set(
-                &mut values_for_color.borrow_mut(),
-                picker.rgba().to_string(),
-            );
-            manager_for_color.preview(&values_for_color.borrow());
-        });
-        field_row.append(&picker);
-        field_row.append(&label);
-        fields.insert(&field_row, -1);
-    }
-    panel.append(&fields);
-    let error = gtk::Label::new(None);
-    error.add_css_class("theme-editor-error");
-    error.set_xalign(0.0);
-    error.set_visible(false);
-    panel.append(&error);
-    let actions = gtk::Box::new(gtk::Orientation::Horizontal, 8);
-    actions.set_halign(gtk::Align::End);
-    let cancel = gtk::Button::with_label("Cancel");
-    cancel.add_css_class("action-dialog-cancel");
-    let save = gtk::Button::with_label("Add theme");
-    save.add_css_class("action-dialog-confirm");
-    actions.append(&cancel);
-    actions.append(&save);
-    panel.append(&actions);
-    let revealer = gtk::Revealer::builder()
-        .transition_type(gtk::RevealerTransitionType::SlideDown)
-        .child(&panel)
-        .build();
-    let hidden = revealer.clone();
-    let manager_for_cancel = manager.clone();
-    cancel.connect_clicked(move |_| {
-        manager_for_cancel.cancel_preview();
-        hidden.set_reveal_child(false);
-    });
-    let hidden = revealer.clone();
-    save.connect_clicked(move |_| {
-        let mut tokens = values.borrow().clone();
-        tokens.name = name.text().trim().to_owned();
-        match manager.save_custom_theme(tokens) {
-            Ok(_) => {
-                error.set_visible(false);
-                hidden.set_reveal_child(false);
-            }
-            Err(message) => {
-                error.set_text(&message.to_string());
-                error.set_visible(true);
-            }
-        }
-    });
-    (revealer, fields)
-}
-
-#[derive(Clone, Copy)]
-enum ColorField {
-    Background,
-    Surface,
-    Text,
-    Accent,
-    Danger,
-    Muted,
-    Highlight,
-    Border,
-    DimText,
-}
-impl ColorField {
-    fn get(self, tokens: &ThemeTokens) -> &str {
-        match self {
-            Self::Background => &tokens.background,
-            Self::Surface => &tokens.surface,
-            Self::Text => &tokens.text,
-            Self::Accent => &tokens.accent,
-            Self::Danger => &tokens.danger,
-            Self::Muted => &tokens.muted,
-            Self::Highlight => &tokens.highlight,
-            Self::Border => &tokens.border,
-            Self::DimText => &tokens.dim_text,
-        }
-    }
-    fn set(self, tokens: &mut ThemeTokens, value: String) {
-        *match self {
-            Self::Background => &mut tokens.background,
-            Self::Surface => &mut tokens.surface,
-            Self::Text => &mut tokens.text,
-            Self::Accent => &mut tokens.accent,
-            Self::Danger => &mut tokens.danger,
-            Self::Muted => &mut tokens.muted,
-            Self::Highlight => &mut tokens.highlight,
-            Self::Border => &mut tokens.border,
-            Self::DimText => &mut tokens.dim_text,
-        } = value;
-    }
-}
-
 fn navigation_button(icon: &str, label: &str) -> (gtk::Button, gtk::Label, gtk::Box) {
     let content = gtk::Box::new(gtk::Orientation::Horizontal, 10);
     let icon_image = crate::assets::primary_icon(icon, 18);
@@ -3188,37 +2654,4 @@ fn append_heading(container: &gtk::Box, text: &str) -> gtk::Label {
     heading.add_css_class("menu-heading");
     container.append(&heading);
     heading
-}
-
-trait RoundedRectangle {
-    fn rounded_rectangle(&self, x: f64, y: f64, width: f64, height: f64, radius: f64);
-}
-impl RoundedRectangle for gtk::cairo::Context {
-    fn rounded_rectangle(&self, x: f64, y: f64, width: f64, height: f64, radius: f64) {
-        let degrees = std::f64::consts::PI / 180.0;
-        self.new_sub_path();
-        self.arc(x + width - radius, y + radius, radius, -90.0 * degrees, 0.0);
-        self.arc(
-            x + width - radius,
-            y + height - radius,
-            radius,
-            0.0,
-            90.0 * degrees,
-        );
-        self.arc(
-            x + radius,
-            y + height - radius,
-            radius,
-            90.0 * degrees,
-            180.0 * degrees,
-        );
-        self.arc(
-            x + radius,
-            y + radius,
-            radius,
-            180.0 * degrees,
-            270.0 * degrees,
-        );
-        self.close_path();
-    }
 }
