+++ src/ui/settings/tests/preferences.rs
@@ -2,6 +2,7 @@
 
 use super::super::*;
 use crate::test_support::gtk_test;
+use crate::ui::theme::TextSize;
 
 fn descendants<T: IsA<gtk::Widget> + Clone>(root: &gtk::Widget) -> Vec<T> {
     let mut widgets = Vec::new();
